+++ backend/canisters/group_index/api/src/updates/mod.rs
@@ -18,6 +18,7 @@ pub mod freeze_community;
 pub mod freeze_group;
 pub mod mark_local_group_index_full;
 pub mod remove_hot_group_exclusion;
+pub mod set_community_moderation_flags;
 pub mod set_community_upgrade_concurrency;
 pub mod set_group_upgrade_concurrency;
 pub mod set_max_concurrent_community_canister_upgrades;